+++ core/src/Kokkos_Core.hpp
@@ -75,6 +75,7 @@
 #include <Kokkos_TaskScheduler.hpp>
 #include <Kokkos_Complex.hpp>
 #include <Kokkos_CopyViews.hpp>
+#include <impl/Kokkos_TeamMDPolicy.hpp>
 #include <impl/Kokkos_InitializationSettings.hpp>
 #include <functional>
 #include <iosfwd>
